requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements/requirements.txt requirements/requirements.in
click==8.3.1
    # via cloup
cloup==3.0.8
    # via -r requirements/requirements.in
colorama==0.4.6
    # via click
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dill==0.4.0
    # via qiskit
fonttools==4.60.1
    # via matplotlib
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.7
    # via -r requirements/requirements.in
mpmath==1.3.0
    # via sympy
numpy==2.3.5
    # via
    #   contourpy
    #   matplotlib
    #   qiskit
    #   qiskit-aer
    #   rustworkx
    #   scipy
packaging==25.0
    # via matplotlib
pillow==12.0.0
    # via
    #   -r requirements/requirements.in
    #   matplotlib
psutil==7.1.3
    # via qiskit-aer
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   qiskit
    #   qiskit-aer
qiskit==1.4.5
    # via
    #   -r requirements/requirements.in
    #   qiskit-aer
qiskit-aer==0.17.2
    # via -r requirements/requirements.in
rustworkx==0.17.1
    # via qiskit
scipy==1.16.3
    # via
    #   qiskit
    #   qiskit-aer
six==1.17.0
    # via python-dateutil
stevedore==5.6.0
    # via qiskit
symengine==0.13.0
    # via qiskit
sympy==1.14.0
    # via qiskit
tabulate==0.9.0
    # via -r requirements/requirements.in
typing-extensions==4.15.0
    # via qiskit
